+++ R/data.R
@@ -12,8 +12,10 @@
 #'   \item{who_code}{\href{https://apps.who.int/gho/data/node.metadata.COUNTRY?lang=en}{World Health Organization code.}}
 #'   \item{m49}{\href{https://en.wikipedia.org/wiki/UN_M49}{United Nations M49 code.}}
 #'   \item{sovereign_iso3}{ISO3 code of sovereign nation. Value is `NA` if sovereign itself.}
-#'   \item{who_member_state}{Logical, is country a WHO member state?}
-#'   \item{who_small_member_state}{Logical, is country a WHO member state with population <= 90,000?}
+#'   \item{who_member}{Logical, is country a WHO member state?}
+#'   \item{who_member_small}{Logical, is country a WHO member state with population <= 90,000?}
+#'   \item{oecd_member}{Logical, is country an OECD member state?}
+#'   \item{gbd_high_income}{Logical, is the country classified as high-income in the IHME's GBD 2019?}
 #'   \item{who_short_name_en}{WHO country name, short, English.}
 #'   \item{who_formal_name_en}{WHO country name, formal, English. Only available for official member states.}
 #'   \item{who_short_name_ar}{WHO country name, short, Arabic.}
@@ -41,6 +43,18 @@
 #'   \item{who_region}{WHO region code.}
 #'   \item{un_region}{UN region code.}
 #'   \item{un_subregion}{UN sub-region code.}
+#'   \item{un_region_name_en}{UN region name, English.}
+#'   \item{un_subregion_name_en}{UN subregion name, English.}
+#'   \item{un_region_name_ru}{UN region name, Russian}
+#'   \item{un_subregion_name_ru}{UN subregion name, Russian}
+#'   \item{un_region_name_fr}{UN region name, French}
+#'   \item{un_subregion_name_fr}{UN subregion name, French}
+#'   \item{un_region_name_es}{UN region name, Spanish.}
+#'   \item{un_subregion_name_es}{UN subregion name, Spanish.}
+#'   \item{un_region_name_ar}{UN region name, Arabic.}
+#'   \item{un_subregion_name_ar}{UN subregion name, Arabic.}
+#'   \item{un_region_name_zh}{UN region name, Chinese.}
+#'   \item{un_subregion_name_zh}{UN subregion name, Chinese.}
 #'   \item{wb_ig_1987}{World Bank Income Group code, 1987 (FY 1989) classification.}
 #'   \item{wb_ig_1988}{World Bank Income Group code, 1988 (FY 1990) classification.}
 #'   \item{wb_ig_1989}{World Bank Income Group code, 1989 (FY 1991) classification.}
@@ -78,5 +92,7 @@
 #' @source \href{https://portal-uat.who.int/xmart4/REFMART}{World Health Organization country reference table, accessed September 23 2020.}
 #' @source \href{https://unstats.un.org/unsd/methodology/m49/overview/}{United Nations Statistics Division, Country or Area Codes for Statistical Use tables; accessed August 15 2020.}
 #' @source \href{https://datahelpdesk.worldbank.org/knowledgebase/articles/906519-world-bank-country-and-lending-groups}{World Bank income groups, fiscal years 1989 - 2021, accessed August 15 2020.}
+#' @source \href{https://www.oecd.org/about/members-and-partners/}{Organisation for Economic Co-operation and Development membership list, accessed 1 February 2021.}
+#' @source \href{http://ghdx.healthdata.org/gbd-2019}{Institute for Health Metrics and Evaluation Global Burden of Disease 2019 Socio-Demographic Index data, accessed 1 February 2021.}
 "countries"